requirements.txt
Flask==3.1.1
flask-cors==5.0.1
python-dotenv==1.1.0
Werkzeug==3.1.3
requests==2.31.0
gunicorn==21.2.0
pycryptodome==3.19.0
Flask-JWT-Extended==4.7.1
paho-mqtt==1.6.1
Flask-SQLAlchemy==3.0.5
psycopg2-binary==2.9.6
flask-mqtt==1.0.3
flask_marshmallow==0.15.0